1. Integrated Neon Serverless Postgres to project. 2. Created Search and Browse Hotels API

+++ src/main/java/com/smit/projects/airBnbApp/repository/InventoryRepository.java
@@ -33,7 +33,7 @@ public interface InventoryRepository extends JpaRepository<Inventory, Long> {
             @Param("startDate") LocalDate startDate,
             @Param("endDate") LocalDate endDate,
             @Param("roomsCount") Integer roomsCount,
-            @Param("dateCount") Integer dateCount,
+            @Param("dateCount") Long dateCount,
             Pageable pageable
             );
